feat: enhance transaksi inputs with new fields and validation, update reports structure, and add utility functions for number formatting

+++ app/Http/Requests/Transaksi/TransaksiInputsRequest.php
@@ -23,7 +23,7 @@ class TransaksiInputsRequest extends FormRequest
         $this->merge(['month' => (int) $this->input('month')]);
         $this->merge([
             'nilais' => array_map(function ($value) {
-                return (int) $value;
+                return (float) $value;
             }, $this->input('nilais', []))
         ]);
     }
